fix: correct import case sensitivity and refactor database schema management
