+++ pyproject.toml
@@ -13,3 +13,6 @@ dependencies = [
     "uvicorn>=0.36.0",
     "webdriver-manager>=4.0.2",
 ]
+
+[tool.uv]
+package = true
