+++ src/scraper/cleaner.py
@@ -1,6 +1,6 @@
 import sqlite3
 import re
-
+from src.database.schema import CREATE_JOBS_TABLE, INSERT_JOBS_TABLE, DELETE_DUPLICATES
 
 def connect_db(db_path="src/data/jobs.db"):
     return sqlite3.connect(db_path)
@@ -34,16 +34,7 @@ def clean_jobs():
     conn = connect_db()
     cursor = conn.cursor()
     try: 
-        cursor.execute(
-            """
-            DELETE FROM jobs
-            WHERE rowid NOT IN (
-                SELECT MIN(rowid)
-                FROM jobs
-                GROUP BY title, company, time, tags, locations, salary, link, logo
-            );
-        """
-        )
+        cursor.execute(DELETE_DUPLICATES)
     except sqlite3.OperationalError:
         pass
 
@@ -53,16 +44,6 @@ def clean_jobs():
         cursor.execute(
             "UPDATE jobs SET tags = ? WHERE rowid = ?", (cleaned_tags, rowid)
         )
-
-    # fields = ["title", "company", "locations"]
-    # for field in fields:
-    #     cursor.execute(
-    #         f"""
-    #         UPDATE jobs
-    #         SET {field} = TRIM(LOWER({field}))
-    #         WHERE {field} IS NOT NULL;
-    #     """
-    #     )
 
     cursor.execute("DELETE FROM jobs WHERE time LIKE '%1yr%'")
 
@@ -83,31 +64,9 @@ def clean_jobs():
     except sqlite3.OperationalError:
         pass
 
-    cursor.execute(
-        """
-        CREATE TABLE IF NOT EXISTS jobs_new (
-            id INTEGER PRIMARY KEY AUTOINCREMENT,
-            title TEXT NOT NULL,
-            company TEXT NOT NULL,
-            time TEXT,
-            tags TEXT,
-            locations TEXT,
-            link TEXT UNIQUE,
-            logo TEXT,
-            salary_from INTEGER,
-            salary_to INTEGER,
-            currency TEXT
-        );
-    """
-    )
+    cursor.execute(CREATE_JOBS_TABLE)
 
-    cursor.execute(
-        """
-        INSERT INTO jobs_new (id, title, company, time, tags, locations, link, logo, salary_from, salary_to, currency)
-        SELECT rowid, title, company, time, tags, locations, link, logo, salary_from, salary_to, currency
-        FROM jobs;
-    """
-    )
+    cursor.execute(INSERT_JOBS_TABLE)
 
     cursor.execute("DROP TABLE jobs;")
     cursor.execute("ALTER TABLE jobs_new RENAME TO jobs;")
